float weights and made gold weigh 0.01f

+++ src/main/java/pw/lemmmy/jrogue/dungeon/items/Item.java
@@ -35,7 +35,7 @@ public abstract class Item {
 
 	public abstract String getName(boolean requiresCapitalisation, boolean plural);
 
-	public abstract int getWeight();
+	public abstract float getWeight();
 
 	public boolean equals(Item other) {
 		return other.getClass() == getClass() &&
